requirements.txt
# openai
groq
langchain
langchain-community
langchain-core
faiss-cpu
langchain-groq
tiktoken
pypdf
streamlit
python-dotenv
sentence-transformers
